requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pypdf
python-docx
pytesseract
Pillow
groq
python-dotenv